requirements.txt
pymysql>=1.0.2
pymongo>=4.6.0
python-dotenv>=1.0.0
tabulate>=0.9.0
